Move the new properties to a extension
