+++ WordPress/MediaEditor/MediaEditor.swift
@@ -40,8 +40,8 @@ public class MediaEditor: NSObject {
 extension MediaEditor: TOCropViewControllerDelegate {
 
     public func cropViewController(_ cropViewController: TOCropViewController, didCropTo image: UIImage, with cropRect: CGRect, angle: Int) {
-        let operations = self.operations(with: cropRect, angle: angle)
-        onFinishEditing?(image, operations)
+        let actions = cropViewController.actions
+        onFinishEditing?(image, actions)
         releaseCallbacks()
     }
 
@@ -49,57 +49,5 @@ extension MediaEditor: TOCropViewControllerDelegate {
         self.cropViewController?.dismiss(animated: true)
         onCancel?()
         releaseCallbacks()
-    }
-
-    private func operations(with cropRect: CGRect, angle: Int) -> [MediaEditorOperation] {
-        var operations: [MediaEditorOperation] = []
-
-        if isCropped {
-            operations.append(.crop)
-        }
-
-        if isRotated {
-            operations.append(.rotate)
-        }
-
-        return operations
-    }
-}
-
-// MARK: - Cropping/Resizing properties
-
-extension MediaEditor {
-    // TOCropViewController sometimes resize the image by 1, 2 or 3 points automatically.
-    // In those cases we're not considering that as a cropping action.
-    var isCropped: Bool {
-        guard let cropViewController = cropViewController else {
-            return false
-        }
-
-        return abs(imageSizeDiscardingRotation.width - cropViewController.image.size.width) > 4 ||
-            abs(imageSizeDiscardingRotation.height - cropViewController.image.size.height) > 4
-    }
-
-    var imageSizeDiscardingRotation: CGSize {
-        guard let cropViewController = cropViewController else {
-            return .zero
-        }
-
-        let imageSize = cropViewController.imageCropFrame.size
-
-        let anglesThatChangesImageSize = [90, 270]
-        if anglesThatChangesImageSize.contains(cropViewController.angle) {
-            return CGSize(width: imageSize.height, height: imageSize.width)
-        } else {
-            return imageSize
-        }
-    }
-
-    var isRotated: Bool {
-        guard let cropViewController = cropViewController else {
-            return false
-        }
-
-        return cropViewController.angle != 0
     }
 }
